requirements.txt
anthropic>=0.40.0
requests>=2.31.0
beautifulsoup4>=4.12.0
lxml>=5.0.0
click>=8.1.0
rich>=13.0.0
python-dotenv>=1.0.0
schedule>=1.2.0
google-auth>=2.27.0
google-auth-oauthlib>=1.2.0
google-api-python-client>=2.115.0
